Splits Person and ManagedPerson code
