+++ WordPress/Classes/Models/Person.swift
@@ -1,10 +1,14 @@
 import Foundation
-import CoreData
 import WordPressShared
 
+// MARK: - Typealiases
+//
 typealias People = [Person]
 
-struct Person {
+
+// MARK: - Person Encapsulates all of the properties a Blog's User may have
+//
+struct Person: Equatable {
     let ID: Int
     let username: String
     let firstName: String?
@@ -23,118 +27,15 @@ struct Person {
         case Author
         case Contributor
         case Unsupported
-
-        init(string: String) {
-            switch string {
-            case "administrator":
-                self = .Admin
-            case "editor":
-                self = .Editor
-            case "author":
-                self = .Author
-            case "contributor":
-                self = .Contributor
-            default:
-                self = .Unsupported
-            }
-        }
-
-        func color() -> UIColor {
-            switch self {
-            case .SuperAdmin:
-                return WPStyleGuide.People.superAdminColor
-            case .Admin:
-                return WPStyleGuide.People.adminColor
-            case .Editor:
-                return WPStyleGuide.People.editorColor
-            case .Author:
-                return WPStyleGuide.People.authorColor
-            case .Contributor:
-                return WPStyleGuide.People.contributorColor
-            case .Unsupported:
-                return WPStyleGuide.People.contributorColor
-            }
-        }
-
-        func localizedName() -> String {
-            switch self {
-            case .SuperAdmin:
-                return NSLocalizedString("Super Admin", comment: "User role badge")
-            case .Admin:
-                return NSLocalizedString("Admin", comment: "User role badge")
-            case .Editor:
-                return NSLocalizedString("Editor", comment: "User role badge")
-            case .Author:
-                return NSLocalizedString("Author", comment: "User role badge")
-            case .Contributor:
-                return NSLocalizedString("Contributor", comment: "User role badge")
-            case .Unsupported:
-                return NSLocalizedString("Unsupported", comment: "User role badge")
-            }
-        }
-
-        var description: String {
-            switch self {
-            case .SuperAdmin:
-                return "super-admin"
-            case .Admin:
-                return "administrator"
-            case .Editor:
-                return "editor"
-            case .Author:
-                return "author"
-            case .Contributor:
-                return "contributor"
-            default:
-                return "unsupported"
-            }
-        }
         
         static let roles : [Role] = [.SuperAdmin, .Admin, .Editor, .Author, .Contributor]
     }
 }
 
-func ==(lhs: Person.Role, rhs: Person.Role) -> Bool {
-    return lhs.rawValue == rhs.rawValue
-}
 
-func <(lhs: Person.Role, rhs: Person.Role) -> Bool {
-    return lhs.rawValue < rhs.rawValue
-}
 
-class ManagedPerson: NSManagedObject {
-    
-}
-
-extension ManagedPerson {
-    @NSManaged var avatarURL: String?
-    @NSManaged var displayName: String
-    @NSManaged var firstName: String?
-    @NSManaged var lastName: String?
-    @NSManaged var role: String
-    @NSManaged var siteID: Int32
-    @NSManaged var userID: Int32
-    @NSManaged var linkedUserID: Int32
-    @NSManaged var username: String
-    @NSManaged var isSuperAdmin: Bool
-    @NSManaged var blog: Blog
-}
-
-extension ManagedPerson {
-    func updateWith(person: Person) {
-        avatarURL = person.avatarURL?.absoluteString
-        displayName = person.displayName
-        firstName = person.firstName
-        lastName = person.lastName
-        role = String(person.role)
-        siteID = Int32(person.siteID)
-        userID = Int32(person.ID)
-        linkedUserID = Int32(person.linkedUserID)
-        username = person.username
-        isSuperAdmin = person.isSuperAdmin
-    }
-}
-
+// MARK: - Person Helper Methods
+//
 extension Person {
     init(managedPerson: ManagedPerson) {
         ID = Int(managedPerson.userID)
@@ -148,11 +49,7 @@ extension Person {
         avatarURL = managedPerson.avatarURL.flatMap { NSURL(string: $0) }
         isSuperAdmin = managedPerson.isSuperAdmin
     }
-}
 
-extension Person: Equatable {}
-
-extension Person {
     var fullName: String {
         let first = firstName ?? String()
         let last = lastName ?? String()
@@ -162,6 +59,82 @@ extension Person {
     }
 }
 
+
+
+// MARK: - Role Helpers
+//
+extension Person.Role {
+    init(string: String) {
+        switch string {
+        case "administrator":
+            self = .Admin
+        case "editor":
+            self = .Editor
+        case "author":
+            self = .Author
+        case "contributor":
+            self = .Contributor
+        default:
+            self = .Unsupported
+        }
+    }
+    
+    func color() -> UIColor {
+        switch self {
+        case .SuperAdmin:
+            return WPStyleGuide.People.superAdminColor
+        case .Admin:
+            return WPStyleGuide.People.adminColor
+        case .Editor:
+            return WPStyleGuide.People.editorColor
+        case .Author:
+            return WPStyleGuide.People.authorColor
+        case .Contributor:
+            return WPStyleGuide.People.contributorColor
+        case .Unsupported:
+            return WPStyleGuide.People.contributorColor
+        }
+    }
+    
+    func localizedName() -> String {
+        switch self {
+        case .SuperAdmin:
+            return NSLocalizedString("Super Admin", comment: "User role badge")
+        case .Admin:
+            return NSLocalizedString("Admin", comment: "User role badge")
+        case .Editor:
+            return NSLocalizedString("Editor", comment: "User role badge")
+        case .Author:
+            return NSLocalizedString("Author", comment: "User role badge")
+        case .Contributor:
+            return NSLocalizedString("Contributor", comment: "User role badge")
+        case .Unsupported:
+            return NSLocalizedString("Unsupported", comment: "User role badge")
+        }
+    }
+    
+    var description: String {
+        switch self {
+        case .SuperAdmin:
+            return "super-admin"
+        case .Admin:
+            return "administrator"
+        case .Editor:
+            return "editor"
+        case .Author:
+            return "author"
+        case .Contributor:
+            return "contributor"
+        default:
+            return "unsupported"
+        }
+    }
+}
+
+
+
+// MARK: - Operator Overloading
+//
 func ==(lhs: Person, rhs: Person) -> Bool {
     return lhs.ID == rhs.ID
         && lhs.username == rhs.username
@@ -173,4 +146,12 @@ func ==(lhs: Person, rhs: Person) -> Bool {
         && lhs.linkedUserID == rhs.linkedUserID
         && lhs.avatarURL == rhs.avatarURL
         && lhs.isSuperAdmin == rhs.isSuperAdmin
+}
+
+func ==(lhs: Person.Role, rhs: Person.Role) -> Bool {
+    return lhs.rawValue == rhs.rawValue
+}
+
+func <(lhs: Person.Role, rhs: Person.Role) -> Bool {
+    return lhs.rawValue < rhs.rawValue
 }
